requirements.txt
aiohttp==3.*
cchardet==2.*
aiodns==2.*
beautifulsoup4==4.*
requests==2.*
pymorphy2==0.8
pymorphy2-dicts==2.*
DAWG-Python==0.7.*
anyio==3.6.2
pytest-asyncio==0.21
